Adicionando Casting

+++ Adivinhacao.c
@@ -14,7 +14,7 @@ int chute;
 
 int tentativas = 1;
 
-int pontos = 1000;
+double pontos = 1000;
 
 
 while (1) {
@@ -54,12 +54,12 @@ while (1) {
     tentativas++;
 
 
-    int pontosperdidos = (chute - numerosecreto) / 2;
+    double pontosperdidos = (chute - numerosecreto) / (double)2;
     pontos = pontos - pontosperdidos;    
 }
 
     printf("Fim de jogo!\n");
     printf("Voce acertou em %d tentativas!\n", tentativas);
-    printf("Total de pontos: %d\n", pontos);
+    printf("Total de pontos: %.1f\n", pontos);
 
 }
